zy 用户管理中的删除操作可以使用，把js文件夹移到web里面，增加了PatientRemoveServlet,在web.xml增加PatientRemoveServlet配置，修改了PatientDaoImpl中删除方法的patients表
